refactor: Move LanguageServer impl from main
