+++ crates/testa-lsp/src/main.rs
@@ -1,19 +1,18 @@
 mod semantic_token;
 mod semantic_token_builder;
+mod document;
+mod language_server;
 
 use std::collections::HashMap;
 use std::sync::Arc;
 
+use testa_core::parser::Parser;
 use tokio::sync::RwLock;
-use tower_lsp::jsonrpc::Result;
 use tower_lsp::lsp_types::*;
-use tower_lsp::{Client, LanguageServer, LspService, Server};
+use tower_lsp::{Client, LspService, Server};
 
-#[derive(Debug)]
-struct Document {
-    text: String,
-    version: i32,
-}
+use crate::document::Document;
+
 
 #[derive(Debug)]
 struct Backend {
@@ -21,90 +20,22 @@ struct Backend {
     documents: Arc<RwLock<HashMap<String, Document>>>,
 }
 
-#[tower_lsp::async_trait]
-impl LanguageServer for Backend {
-    async fn initialize(&self, _: InitializeParams) -> Result<InitializeResult> {
-        Ok(InitializeResult {
-            capabilities: ServerCapabilities {
-                text_document_sync: Some(TextDocumentSyncCapability::Kind(
-                    TextDocumentSyncKind::FULL,
-                )),
-                semantic_tokens_provider: Some(
-                    SemanticTokensServerCapabilities::SemanticTokensOptions(
-                        SemanticTokensOptions {
-                            work_done_progress_options: WorkDoneProgressOptions::default(),
-                            legend: SemanticTokensLegend {
-                                token_types: vec![
-                                    SemanticTokenType::KEYWORD,
-                                    SemanticTokenType::TYPE,
-                                    SemanticTokenType::ENUM,
-                                    SemanticTokenType::ENUM_MEMBER,
-                                    SemanticTokenType::PROPERTY,
-                                    SemanticTokenType::STRING,
-                                    SemanticTokenType::NUMBER,
-                                    SemanticTokenType::OPERATOR,
-                                    SemanticTokenType::COMMENT,
-                                    SemanticTokenType::MACRO,
-                                    SemanticTokenType::DECORATOR,
-                                ],
-                                token_modifiers: vec![
-                                    SemanticTokenModifier::DECLARATION,
-                                    SemanticTokenModifier::DEFINITION,
-                                    SemanticTokenModifier::READONLY,
-                                    SemanticTokenModifier::DEPRECATED,
-                                ],
-                            },
-                            range: Some(true),
-                            full: Some(SemanticTokensFullOptions::Bool(true)),
-                        },
-                    ),
-                ),
-                ..Default::default()
-            },
-            ..Default::default()
-        })
-    }
-
-    async fn did_open(&self, params: DidOpenTextDocumentParams) {
-        let uri = params.text_document.uri.to_string();
-        let text = params.text_document.text.clone();
-
-        self.documents.write().await.insert(
-            uri,
-            Document {
-                text,
-                version: params.text_document.version,
-            },
-        );
-    }
-
-    async fn did_change(&self, params: DidChangeTextDocumentParams) {
-        let uri = params.text_document.uri.to_string();
-        let text = params.content_changes[0].text.clone();
-
-        if let Some(doc) = self.documents.write().await.get_mut(&uri) {
-            doc.text = text;
-            doc.version = params.text_document.version;
+impl Backend {
+    pub fn new(client: Client) -> Self {
+        Self {
+            client,
+            documents: Arc::new(RwLock::new(HashMap::new())),
         }
     }
 
-    async fn semantic_tokens_full(
-        &self,
-        params: SemanticTokensParams,
-    ) -> Result<Option<SemanticTokensResult>> {
-        let _uri = params.text_document.uri.to_string();
-
-        Ok(None)
-    }
-
-    async fn initialized(&self, _: InitializedParams) {
+    async fn update_document(&self, uri: Url, text: String, _version: i32) {
         self.client
-            .log_message(MessageType::INFO, "server initialized!")
+            .log_message(MessageType::INFO, format!("Parsing document: {}", uri))
             .await;
-    }
-
-    async fn shutdown(&self) -> Result<()> {
-        Ok(())
+        
+        let mut _parser = Parser::new(&text);
+        // self.documents.insert(uri.clone(), document);
+        // self.publish_diagnostics(&uri, diagnostics).await;
     }
 }
 
@@ -113,9 +44,6 @@ async fn main() {
     let stdin = tokio::io::stdin();
     let stdout = tokio::io::stdout();
 
-    let (service, socket) = LspService::new(|client| Backend {
-        client,
-        documents: Arc::new(RwLock::new(HashMap::new())),
-    });
+    let (service, socket) = LspService::new(Backend::new);
     Server::new(stdin, stdout, socket).serve(service).await;
 }
